requirements.txt
langchain
langchain-openai
langgraph
python-dotenv
scikit-learn
colorama
